- Begin implementation of trust/score caching using LRUCache - Fix variable names in trust-related functions: "treeOwner"/"target" should be "truster"/"trustee"... the mistakes are a result of copy-pasting from score-related functions - Trust values are byte, not integer... correct that.

+++ src/plugins/Freetalk/WoT/WoTOwnIdentity.java
@@ -66,7 +66,7 @@ public final class WoTOwnIdentity extends WoTIdentity implements OwnIdentity {
 		return mFreetalk.getIdentityManager().getTrust(this, identity);
 	}
 
-	public void setTrust(WoTIdentity identity, int trust, String comment) throws Exception {
+	public void setTrust(WoTIdentity identity, byte trust, String comment) throws Exception {
 		mFreetalk.getIdentityManager().setTrust(this, identity, trust, comment);
 	}
